move private stuff into src
