+++ src/_exec-browser.js
