+++ test/api-test.js
@@ -1,5 +1,5 @@
 var test = require('tape')
-var exec = require('../_exec')
+var exec = require('../src/_exec')
 
 test('sanity', t=> {
   t.plan(1)
